requirements.txt
asgiref==3.3.1
Django==3.1.5
django-tinymce4-lite==1.8.0
jsmin==2.2.2
Pillow==8.1.0
pytz==2020.5
sqlparse==0.4.1
